add optional loading of machine-local .zshenv
Motivation: I host an MPD server on one machine and connect to it with
many clients, and stream the music from there over HTTP. When the client
is on the same LAN (or the same machine!) as the server, then the
value to set in MPD_HOST differs. So it's best to use a machine-local
configuration for that. Not to mention that the MPD_HOST variable
contains the password for the MPD server, and should not be stored in
version control :)