add `MAGNET_DIR` environment variable to zprofile
The `magnet.sh` script for converting magnet links into torrents stored
the resulting torrents in the directory given by `MAGNET_DIR` or `HOME`
if `MAGNET_DIR` is undefined. On this system, rtorrent is set up to
watch for torrents in `$HOME/torrents/.watch`, so `MAGNET_DIR` should be
set to that path.