comment out the compiler directive for .hs files in .vimrc
When I was using the vim2hs plugin, this integrated GHC with vim.
However, I didn't really like vim2hs, so I stopped using it, but had
forgotten to remove this line from the vimrc. This causes annoying error
messages every time vim entered a buffer for a haskell file.