xmonad: improve mpd integration
Rather than invoke mpc directly, we add a layer of indirection via the scripts. This allows us to modify the scripts to change the behaviour of the hotkeys rather than have to hack the xmonad configuration, which is annoying because testing them requires restarting xmonad in that case.

Also, we add hotkeys for the `-remote` mpd scripts, which controls the MPD
instance on the labcoders box.