xmonad: eliminate hardcoded screenwidth
Now xmonad figures out which is the leftmost screen when it starts and
spawns the two dzen toolbars on it appropriately.

This was a somewhat complex change because we need to be in the X
monad to figure out screen information. But we need this information
to spawn the dzen bars, which we used to do in main, *prior* to even
starting xmonad.

To resolve this, we proxy the logHook through a Chan: rather than
write directly to a Handle, the logHook's ppOutput writes to a Chan
whose consumer is set up inside myStartupHook.
In myStartupHook, we figure out the leftmost screen width and launch
the toolbars. We launch the toolbar that services the loghook in a
separate thread and pump the Chan for new lines and write them into
the Handle for the toolbar as we receive them.