vim: opening external files via xdg-open on gx
Trying to use vim as a 'browser' by maintaining index files with links,
in particular in my journal repo. This works pretty well as long as the
files can be opened by vim, e.g. markdown files. However, when referring
to PDFs, which happens often enough for work documents and research
papers, there needs to be a way to open the file with an external
viewer. Doing this thru xdg-open is create because it isolates the
external opening logic from vim.

It doesn't _totally_ work, unfortunately.